requirements.txt
ollama
chromadb
langgraph
langgraph-sdk
langgraph-checkpoint-sqlite
langsmith
langchain-community
langchain-core
langchain-ollama
tavily-python
python-dotenv
pyyaml
jinja2
pydantic
pydantic[email]
pandas
setuptools
importlib
pathlib
pytest
